moved FileHandler classes from core/session.py to new module io/session.py
